refactor: extract OpenAPI assistant logic into a reusable hook
